api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20
